pyproject.toml
[project]
name = "deepseek-beta"
version = "0.1.0"
description = ""
authors = [
    {name = "MotiTheWizerd",email = "motielmakayes@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "google-adk (>=0.4.0,<0.5.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-ai (>=0.1.9,<0.2.0)",
    "crewai-tools (>=0.43.0,<0.44.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "eel (>=0.18.1,<0.19.0)",
    "fastapi (>=0.110.0,<1.0.0)",
    "uvicorn (>=0.27.0,<1.0.0)",
    "crawl4ai (>=0.6.2,<0.7.0)",
    "playwright (>=1.52.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
